test: colocate split module tests
